+++ crates/block2/src/block.rs
@@ -3,7 +3,7 @@ use core::mem;
 
 use objc2::encode::{EncodeArgument, EncodeReturn, Encoding, RefEncode};
 
-use crate::ffi;
+use crate::abi;
 
 /// Types that may be used as the arguments of an Objective-C block.
 ///
@@ -92,7 +92,7 @@ pub struct Block<A, R> {
     // We store `Block_layout` + a bit more, but `Block` has to remain an
     // empty type otherwise the compiler thinks we only have provenance over
     // `Block_layout`.
-    _layout: PhantomData<ffi::Block_layout>,
+    _layout: PhantomData<abi::Block_layout>,
     // To get correct variance on args and return types
     _p: PhantomData<fn(A) -> R>,
 }
@@ -113,7 +113,7 @@ impl<A: BlockArguments, R: EncodeReturn> Block<A, R> {
     /// caller must ensure that calling it will not cause a data race.
     pub unsafe fn call(&self, args: A) -> R {
         let ptr: *const Self = self;
-        let layout = unsafe { ptr.cast::<ffi::Block_layout>().as_ref().unwrap_unchecked() };
+        let layout = unsafe { ptr.cast::<abi::Block_layout>().as_ref().unwrap_unchecked() };
         // TODO: Is `invoke` actually ever null?
         let invoke = layout.invoke.unwrap_or_else(|| unreachable!());