+++ crates/block2/src/concrete_block.rs
@@ -7,7 +7,7 @@ use std::os::raw::c_ulong;
 
 use objc2::encode::{EncodeArgument, EncodeReturn, Encoding, RefEncode};
 
-use crate::{ffi, Block, BlockArguments, RcBlock};
+use crate::{abi, ffi, Block, BlockArguments, RcBlock};
 
 mod private {
     pub trait Sealed<A> {}
@@ -163,7 +163,7 @@ concrete_block_impl!(
 #[repr(C)]
 pub struct ConcreteBlock<A, R, F> {
     p: PhantomData<Block<A, R>>,
-    pub(crate) layout: ffi::Block_layout,
+    pub(crate) layout: abi::Block_layout,
     pub(crate) closure: F,
 }
 
@@ -187,14 +187,14 @@ where
 
 impl<A, R, F> ConcreteBlock<A, R, F> {
     // TODO: Use new ABI with BLOCK_HAS_SIGNATURE
-    const FLAGS: ffi::block_flags = if mem::needs_drop::<Self>() {
-        ffi::BLOCK_HAS_COPY_DISPOSE
+    const FLAGS: abi::block_flags = if mem::needs_drop::<Self>() {
+        abi::BLOCK_HAS_COPY_DISPOSE
     } else {
         0
     };
 
-    const DESCRIPTOR: ffi::Block_descriptor = ffi::Block_descriptor {
-        header: ffi::Block_descriptor_header {
+    const DESCRIPTOR: abi::Block_descriptor = abi::Block_descriptor {
+        header: abi::Block_descriptor_header {
             reserved: 0,
             size: mem::size_of::<Self>() as c_ulong,
         },
@@ -214,12 +214,12 @@ impl<A, R, F> ConcreteBlock<A, R, F> {
     /// Unsafe because the caller must ensure the invoke function takes the
     /// correct arguments.
     unsafe fn with_invoke(invoke: unsafe extern "C" fn(), closure: F) -> Self {
-        let layout = ffi::Block_layout {
+        let layout = abi::Block_layout {
             isa: unsafe { &ffi::_NSConcreteStackBlock },
             flags: Self::FLAGS,
             reserved: 0,
             invoke: Some(invoke),
-            descriptor: &Self::DESCRIPTOR as *const ffi::Block_descriptor as *mut c_void,
+            descriptor: &Self::DESCRIPTOR as *const abi::Block_descriptor as *mut c_void,
         };
         Self {
             p: PhantomData,