+++ crates/block2/src/ffi.rs
@@ -1,0 +1,6 @@
+//! `Block.h`
+
+pub use block_sys::{
+    Class, _Block_copy, _Block_object_assign, _Block_object_dispose, _Block_release,
+    _NSConcreteGlobalBlock, _NSConcreteStackBlock,
+};